+++ test/util/decimal_test.dart
@@ -5,10 +5,13 @@ import 'package:folly_fields/util/decimal.dart';
 ///
 ///
 void main() {
-  test('Check decimal values.', () {
-    for (int integer = 0; integer <= 9999999; integer++) {
-      Decimal decimal = Decimal(precision: 2, intValue: integer);
-      expect(decimal.intValue, integer);
-    }
-  });
+  test(
+    'Check decimal values.',
+    () {
+      for (int integer = 0; integer <= 9999999; integer++) {
+        Decimal decimal = Decimal(precision: 2, intValue: integer);
+        expect(decimal.intValue, integer);
+      }
+    },
+  );
 }